backend/python/services/app/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
PyJWT[crypto]
psycopg[binary]
